Pull the decryption key out of task.shtml
(Hopefully the surrounding spaces aren't actually part of the key?)

I should actually have a key of my own on an old hard drive somewhere,
but it would be a bit of a bother to dig it out and doesn't seem all
that relevant since they froze the scoreboard after the contest.

Note: Originally, each team recieved their own key, which was
essentially used to lock the "publication" strings to that team.

At the time, these "publication" strings could be used to "prove" that
you had solved certain puzzles both in-Codex and on the contest website.
Each "publication" is worth a set number of points, and when submitted
to the website, the points would be tallied to the appropriate team
automatically. But that's not really relevant now, so it doesn't really
matter whose key you use anymore, and thankfully boundvariable.org
provides a pre-generated key now.

I'm pretty sure that pasting enough points' worth of publications into
the in-Codex verifier can be used to unlock more puzzles somehow, but I
forget exactly how it works and I haven't taken so many shortcuts in
writing my UM this time around so I can't check yet.

(Last time, I was on a 32-bit machine, I wrote it in C, and I left out
support for replacing array 0 in my main implementation, so I could use
array-0-relative addresses for the array identifiers and it was
extremely straightforward to implement each instruction. This was not
super efficient, but it wasn't super slow either, and it was very easy
to write, and I think it faithfully implemented the semantics of
everything except "Load Program" with a non-zero array index, which
the codex apparently doesn't actually use.)